source/DBS_database/requirements.txt
Django
psycopg2-binary
python-dateutil